requirements.txt
docker-fabric
fabric
cuisine==0.7.11
enum
pyrax
boto
moto
ipython
oauth2client
google-api-python-client
zope.interface
pyrsistent
# paramiko doesn't support recent SSH macs
# https://github.com/paramiko/paramiko/pull/581
-e git+https://github.com/ericwb/paramiko.git@rfc6668#egg=paramiko
